attempt to fix inf2cat related issues
There is an issue related to the inf2cat (bundled with EWDK ISO) where it reports a 'File not found' like error when attempting to look for the *.sys file generated by the previous step. Changes from this commit will be used in a seperate testing workstation to see if the issue persists.